Isolated long-running *Store and *Graph tests
fix not found failures

Added a heartbeat timer. This heartbeat timer will print to the console every x configured milliseconds. This is to prevent Travis CI from terminating the build if there is no console output in a 10min time frame. The heartbeat will only run for up to x configured intervals, this is needed for tests which are really in a hung state and Travis CI can then fail the build.

The heartbeat is started and stopped before each test case. the current running unit test and the current time execution of the test is included in the heartbeat for better visibility in to any long running tests.

Disabled metrics printing to console

Extended the default iteration count for the heartbeat. Moved the heartbeat to all abstract classes

Fixed compile error

Extended the default iteration count for the heartbeat.

Uncommented all of the modules in the Travis config file

Extended the default iteration count for the heartbeat.

Added output for unit test start and finish events

Set the correct email address again in the Travis config file