Better handling of failed downloads (#207)
Introduced a command line option "-f" for install-gremlin-server.sh that
forces re-downloading of the files and ignores the cache contents.

Rolled back to the original 1.2.1 version of the download plugin - looks
like the overwrite option does no behave as documented in 1.3.0.

Fixed minor issue with one of mkdir commands, the script was failing if
started for the second time.